Extract indentation code out of main into a separate module
The new function accepts a vector containing lines of code. The main
function reads lines from stdin and constructs this vector then passed
it (as a value) to the 'hard_indent_org' function.

This is inefficient in terms of memory utilization as we're reading a
lines of a file into memory. There is a scope for optimization by
using some kind of a buffer or a lazy iterator constructed from
stdin. Planning to tackle this later after a better understanding of
laziness and lifetimes in rust.