Add structure and handle errors in cli processing
Defined an 'execute' method for the `Cli` struct which actually
matches the args and runs the code. It returns a result type. For now,
we're returning a tuple as part of 'Ok' and a '&str' as part of 'Err'.

The 'main' function calls the 'execute' method and based on the
returned value, it either exits with exit code 0 (success) or
1 (failure). In case of failure, the error string is printed to stderr
using the 'eprintln!' macro.